fix ball speed issues when colliding with the bat
To keep the direction as unit vector we need to normalize it after
making any changes to it, Otherwise, the speed component of the velocity
will change. Thus, added an instruction to normalize direction_ after
collision with the bat. Remeber: we change direction_ depending on the
angle the ball hits the bat.